brick breaker game
hey, I was following a tutorial on youtube, when i got to 7:36 of https://www.youtube.com/watch?v=9AzSlN8ADFE this is when I ran into trouble. for the life of me I cant find a difference between the code that works in the video and the code of my screen that doesnt. why I run the program it crashes immediately, when I run this in the python console it reads 
"""
Traceback (most recent call last):
  File "C:\Program Files\JetBrains\PyCharm Community Edition 2022.3.2\plugins\python-ce\helpers\pydev\pydevconsole.py", line 364, in runcode
    coro = func()
  File "<input>", line 1, in <module>
  File "C:\Program Files\JetBrains\PyCharm Community Edition 2022.3.2\plugins\python-ce\helpers\pydev\_pydev_bundle\pydev_import_hook.py", line 21, in do_import
    module = self._system_import(name, *args, **kwargs)
ModuleNotFoundError: No module named 'pygame'
"""

I came down to finding that if i comment out line 123 "player_paddle.draw()" the game runs, and when I uncomment that line it breaks, 


I have checked character for character, "," for ",", indentations, spelling. I have no idea why the program isnt running the same as the one in the video. 


yes I've installed pygame, and even followed a different video to make a different game work